[SPARK-19799][SQL] Support WITH clause in subqueries
